backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
torch
torchvision
ultralytics
opencv-python-headless